Chore: shift constants file into separate folder
