+++ src/contexts/LoginContext.jsx
@@ -1,7 +1,7 @@
 import axios from "axios"
 import React, { createContext, useEffect, useState, useContext } from "react"
 
-import { SITES_IS_PRIVATE_KEY } from "../constants"
+import { SITES_IS_PRIVATE_KEY } from "constants/constants"
 
 const { REACT_APP_BACKEND_URL: BACKEND_URL } = process.env
 const LOCAL_STORAGE_USER_ID_KEY = "userId"
